requirements.txt
streamlit
opencv-python
tensorflow
pymysql
numpy
pillow
